Condensed packaged and hotel deals service and added hotel deals json file
