+++ src/app/services/deals/adeal.ts
@@ -1,3 +1,4 @@
 export abstract class ADeal {
   public abstract async getDealList(): Promise<any>;
+  public abstract async getHotelDealList(): Promise<any>;
 }
