+++ src/app/services/deals/deal-api.service.spec.ts
