+++ src/assets/mock-json-tests/hotel-deals.json
@@ -1,0 +1,8 @@
+{
+  "deals": [{
+   "name": "B Akasaka",
+   "image": "B aksasaka Image",
+   "percentOff": "30",
+   "price": 67
+  }]
+}
